+++ config/webpack/environment.js
@@ -1,5 +1,6 @@
 const { environment } = require('@rails/webpacker')
 const webpack = require('webpack')
+const CompressionPlugin = require('compression-webpack-plugin')
 
 environment.plugins.prepend(
   'Provide',
@@ -7,6 +8,14 @@ environment.plugins.prepend(
     $: 'jquery',
     jQuery: 'jquery',
     jquery: 'jquery'
+  }),
+  new CompressionPlugin({
+    filename: '[path].gz[query]',
+    algorithm: 'brotliCompress',
+    test: /\.js$|\.css$/,
+    threshold: 10240,
+    minRatio: 0.8,
+    cache: true
   })
 )