+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Modified
 - decrease amount of css
 - convert png icons to webp format
+- add compression for webpacker
 
 ### Fixed
 - creating/updating events